+++ nomos-services/network/Cargo.toml
@@ -19,9 +19,9 @@ tracing        = "0.1"
 utoipa         = { version = "4.0", optional = true }
 
 [dev-dependencies]
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1"
-tracing-subscriber = "0.3"
+tokio              = { version = "1", features = ["full"] }
+tracing            = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 [features]
 default = []
