+++ nomos-services/network/src/backends/libp2p/command.rs
@@ -1,37 +1,8 @@
-use std::collections::HashMap;
-
-use nomos_libp2p::{libp2p::kad::PeerInfo, Multiaddr, PeerId};
+use nomos_libp2p::{Multiaddr, PeerId};
 use serde::{Deserialize, Serialize};
 use tokio::sync::oneshot;
 
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum PubSubCommand {
-    Broadcast {
-        topic: Topic,
-        message: Box<[u8]>,
-    },
-    Subscribe(Topic),
-    Unsubscribe(Topic),
-    #[doc(hidden)]
-    RetryBroadcast {
-        topic: Topic,
-        message: Box<[u8]>,
-        retry_count: usize,
-    },
-}
-
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum DiscoveryCommand {
-    GetClosestPeers {
-        peer_id: PeerId,
-        reply: oneshot::Sender<Vec<PeerInfo>>,
-    },
-    DumpRoutingTable {
-        reply: oneshot::Sender<HashMap<u32, Vec<PeerId>>>,
-    },
-}
+use crate::backends::libp2p::swarm::{DiscoveryCommand, PubSubCommand};
 
 #[derive(Debug)]
 #[non_exhaustive]
@@ -54,8 +25,6 @@ pub struct Dial {
     pub retry_count: usize,
     pub result_sender: oneshot::Sender<Result<PeerId, nomos_libp2p::DialError>>,
 }
-
-pub type Topic = String;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
